foundation-project/pyproject.toml
[project]
name = "foundation-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "ipykernel>=7.2.0",
    "openai>=2.21.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "typing>=3.10.0.0",
]
